// projects/ai-gantt-chart/src/lib/utils/enums.ts
export enum ViewMode {
  QuarterDay = 'Quarter Day',
  HalfDay = 'Half Day',
  Day = 'Day',
  Week = 'Week',
  Month = 'Month',
  Year = 'Year'
}

export enum Language {
  English = 'en',
  Spanish = 'es',
  Russian = 'ru',
  Portuguese = 'ptBr',
  French = 'fr'
}

export enum Options {
  PopupTrigger = 'click',
  DateFormat = 'YYYY-MM-DD',
}

export enum Scale {
  Year = 'year',
  Month = 'month',
  Day = 'day',
  Hour = 'hour',
  Minute = 'minute',
  Second = 'second',
  Millisecond = 'millisecond'
}